Move all the closely tied to GHC modules together
