+++ compiler/hie-core/hie-core.cabal
@@ -77,20 +77,20 @@ library
         ../../libs-haskell/prettyprinter-syntax/src
     exposed-modules:
         Development.IDE.Logger
-        Development.IDE.UtilGHC
+        Development.IDE.GHC.Util
         Development.IDE.Spans.AtPoint
         Development.IDE.Functions.Compile
-        Development.IDE.Functions.CPP
-        Development.IDE.Orphans
+        Development.IDE.GHC.CPP
+        Development.IDE.GHC.Orphans
         Development.IDE.Import.DependencyInformation
         Development.IDE.Spans.Documentation
         Development.IDE.Import.FindImports
-        Development.IDE.Functions.GHCError
+        Development.IDE.GHC.Error
         Development.IDE.Spans.Calculate
-        Development.IDE.Functions.Warnings
+        Development.IDE.GHC.Warnings
         Development.IDE.State.FileStore
         Development.IDE.State.Rules
-        Development.IDE.Compat
+        Development.IDE.GHC.Compat
         Development.IDE.LSP.LanguageServer
         Development.IDE.LSP.Definition
         Development.IDE.LSP.Hover
